Move RealmTest.java to the new testsuite
Part of: #34494
