+++ tests/base/src/test/resources/org/keycloak/tests/admin/realm/saml-entity-descriptor.xml
