+++ tests/base/src/test/resources/org/keycloak/tests/admin/realm/testrealm-keycloak-6146-error.json
