+++ tests/base/src/test/resources/org/keycloak/tests/admin/realm/testrealm.json
